fix: update 쿼리 리턴값 수정
update 쿼리는 업데이트 성공한 행 수를 Int로 반환함.
따라서 Long -> Int로 수정

+++ app/src/main/java/com/example/notimanager/data/repository/NotificationAppRepository.kt
@@ -26,13 +26,13 @@ class NotificationAppRepository(
     override suspend fun setAppPriority(
         appName: String,
         newPriority: Int
-    ): Long {
+    ): Int {
         return appIconDao.setPriority(appName, newPriority)
     }
 
     override suspend fun removeAppPriority(
         appName: String,
-    ): Long {
+    ): Int {
         return appIconDao.removePriority(appName)
     }
 }
