+++ app/src/main/java/com/example/notimanager/data/source/local/dao/NotificationIconDao.kt
@@ -12,8 +12,8 @@ interface NotificationIconDao {
     suspend fun insert(notificationIconModel: NotificationIconModel): Long
 
     @Query("UPDATE notification_icon SET priorityActive = 1, priority = :newPriority WHERE notificationId = :notificationId")
-    suspend fun setPriority(notificationId: Long, newPriority: Int): Long
+    suspend fun setPriority(notificationId: Long, newPriority: Int): Int
 
     @Query("UPDATE notification_icon SET priorityActive = 0, priority = 0 WHERE notificationId = :notificationId")
-    suspend fun removePriority(notificationId: Long): Long
+    suspend fun removePriority(notificationId: Long): Int
 }
