+++ app/src/main/java/com/example/notimanager/domain/repository/NotificationAppRepositoryInterface.kt
@@ -5,6 +5,6 @@ import com.example.notimanager.domain.model.NotificationApp
 interface NotificationAppRepositoryInterface {
     suspend fun getNotificationAppList(): List<NotificationApp>
     suspend fun getNotificationAppPriorityList(): List<NotificationApp>
-    suspend fun setAppPriority(appName: String, newPriority: Int): Long
-    suspend fun removeAppPriority(appName: String): Long
+    suspend fun setAppPriority(appName: String, newPriority: Int): Int
+    suspend fun removeAppPriority(appName: String): Int
 }
